+++ fiddles/bash/fiddle-0093-InitAndroidProject/scripts/bin/android/_create.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+function createAndroid() {
+    android create project -a Main -k com.example.app -t 19 -g -v 0.10 -p AppWithGradleTemplate
+
+}